Move ir::Type to ast::Type (replacing Datatypes.h with Type.h)
