+++ odb-compiler/include/odb-compiler/ast/CommandExpr.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 namespace odb::cmd {
-    class Command;
+class Command;
 }
 
 namespace odb::ast {
