+++ odb-compiler/include/odb-compiler/ast/Literal.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "odb-compiler/ast/Expression.hpp"
-#include "odb-compiler/ast/Datatypes.hpp"
+#include "odb-compiler/ast/Type.hpp"
 
 namespace odb::ast {
