+++ odb-compiler/include/odb-compiler/ast/Loop.hpp
@@ -4,8 +4,7 @@
 #include "odb-compiler/ast/Statement.hpp"
 #include "odb-sdk/MaybeNull.hpp"
 
-namespace odb {
-namespace ast {
+namespace odb::ast {
 
 class AnnotatedSymbol;
 class Block;
@@ -116,5 +115,4 @@ private:
     Reference<Block> body_;
 };
 
-}
 }
