+++ odb-compiler/include/odb-compiler/ast/Statement.hpp
@@ -3,8 +3,7 @@
 #include "odb-compiler/config.hpp"
 #include "odb-compiler/ast/Node.hpp"
 
-namespace odb {
-namespace ast {
+namespace odb::ast {
 
 /* A single executable statement */
 class ODBCOMPILER_PUBLIC_API Statement : public Node
@@ -13,5 +12,4 @@ public:
     Statement(SourceLocation* location);
 };
 
-}
 }
