+++ odb-compiler/include/odb-compiler/ast/Visitor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "odb-compiler/config.hpp"
-#include "odb-compiler/ast/Datatypes.hpp"
+#include "odb-compiler/ast/Type.hpp"
 
 namespace odb::ast {
