+++ odb-compiler/src/ast/Expression.cpp
@@ -1,8 +1,7 @@
 #include "odb-compiler/ast/Expression.hpp"
 #include "odb-compiler/ast/SourceLocation.hpp"
 
-namespace odb {
-namespace ast {
+namespace odb::ast {
 
 // ----------------------------------------------------------------------------
 Expression::Expression(SourceLocation* location) :
@@ -10,5 +9,4 @@ Expression::Expression(SourceLocation* location) :
 {
 }
 
-}
 }
