+++ odb-compiler/src/ast/Node.cpp
@@ -4,11 +4,11 @@
 #include "odb-compiler/commands/Command.hpp"
 #include "odb-compiler/parsers/db/Parser.y.hpp"
 
-namespace odb {
-namespace ast {
+namespace odb::ast {
 
 // ----------------------------------------------------------------------------
 Node::Node(SourceLocation* location) :
+    parent_(nullptr),
     location_(location)
 {
 }
@@ -25,5 +25,4 @@ SourceLocation* Node::location() const
     return location_;
 }
 
-}
 }
