+++ odb-compiler/src/ir/semantic/ASTConverter.cpp
@@ -15,7 +15,6 @@
 #include "odb-compiler/ast/Conditional.hpp"
 #include "odb-compiler/ast/ConstDecl.hpp"
 #include "odb-compiler/ast/Data.hpp"
-#include "odb-compiler/ast/Datatypes.hpp"
 #include "odb-compiler/ast/Exit.hpp"
 #include "odb-compiler/ast/Exporters.hpp"
 #include "odb-compiler/ast/Expression.hpp"
@@ -31,6 +30,7 @@
 #include "odb-compiler/ast/Statement.hpp"
 #include "odb-compiler/ast/Subroutine.hpp"
 #include "odb-compiler/ast/Symbol.hpp"
+#include "odb-compiler/ast/Type.hpp"
 #include "odb-compiler/ast/UDTDecl.hpp"
 #include "odb-compiler/ast/UDTRef.hpp"
 #include "odb-compiler/ast/UnaryOp.hpp"
